+++ unit_test/unit_test.F90
@@ -1,7 +1,8 @@
-subroutine init_runtime_parameters(name, namlen) bind(C, name="init_runtime_parameters")
+subroutine init_unit_test_fortran(name, namlen) bind(C, name="init_unit_test_fortran")
 
   use amrex_fort_module, only: rt => amrex_real
   use extern_probin_module
+  use microphysics_module
 
   implicit none
 
@@ -10,16 +11,16 @@ subroutine init_runtime_parameters(name, namlen) bind(C, name="init_runtime_para
 
   call runtime_init(name, namlen)
 
-end subroutine init_runtime_parameters
+end subroutine init_unit_test_fortran
 
-subroutine init_fortran_microphysics() bind(C, name="init_fortran_microphysics")
+subroutine init_microphysics_fortran() bind(C, name="init_microphysics_fortran")
 
   use amrex_fort_module, only: rt => amrex_real
+  use extern_probin_module
   use microphysics_module
-  use extern_probin_module, only : small_temp, small_dens
 
   implicit none
 
   call microphysics_init(small_temp, small_dens)
 
-end subroutine init_fortran_microphysics
+end subroutine init_microphysics_fortran