+++ CHANGES.md
@@ -3,6 +3,11 @@
    * simplified-SDC can be used with the NSE table in aprox19 now
      (#423, #497)
 
+   * Runtime parameters can now be set in the inputs file instead of
+     the probin file (and then they are read in by C++ ParmParse).  If
+     a parameter is set in both places, then the inputs value is used.
+     (#505)
+
    * Fortran support for simplified-SDC in the VODE integrator has
      been removed. (#492)